Refactored Config and Changed how REST call works
Config was storing values for a quick use. Now I have the code use the
config object as it was intended.
I changed the REST call for mentions to pick a mention and update config
with its id. I'm currently working on determing the user that made the
mention. I also need to determine if the id I'm storing is correct.